+++ provision/pack.go
@@ -12,6 +12,7 @@
 package provision
 
 import (
+	"bufio"
 	"crypto/md5"
 	"crypto/rand"
 	"encoding/hex"
@@ -21,6 +22,7 @@ import (
 	"os"
 	"os/exec"
 	"path"
+	"regexp"
 	"strings"
 	"sync"
 	"time"
@@ -30,7 +32,10 @@ import (
 	"gopkg.in/inconshreveable/log15.v2"
 )
 
-var dockerSaveLock sync.Mutex
+var (
+	dockerSaveLock sync.Mutex
+	buildArgRe     = regexp.MustCompile(`^ARG (.+)$`)
+)
 
 // Package creates the service payload to deliver to S3
 func Package(log log15.Logger, config *conf.Config) (success bool) {
@@ -117,8 +122,7 @@ func Package(log log15.Logger, config *conf.Config) (success bool) {
 
 		go func(container *conf.Container) {
 
-			successChan <- buildContainer(log, container.Name,
-				container.Dockerfile, container.DockerfileBuild)
+			successChan <- buildContainer(log, container)
 
 		}(container)
 	}
@@ -172,28 +176,71 @@ func Package(log log15.Logger, config *conf.Config) (success bool) {
 	return
 }
 
-func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild string) (success bool) {
+func buildContainer(log log15.Logger, container *conf.Container) (success bool) {
 
-	log = log.New("ImageTag", containerName)
+	log = log.New("ImageTag", container.Name)
 
-	imagePath := fmt.Sprintf("%s/%s.docker", constants.PayloadWorkingDir, containerName)
+	imagePath := fmt.Sprintf("%s/%s.docker", constants.PayloadWorkingDir, container.Name)
 
-	_, err := os.Stat(dockerfile)
+	_, err := os.Stat(container.Dockerfile)
 	if err != nil {
 		log.Error("Dockerfile stat", "Error", err)
 		return
 	}
 
 	haveBuilder := true
-	_, err = os.Stat(dockerfileBuild)
+	_, err = os.Stat(container.DockerfileBuild)
 	if err != nil {
 		haveBuilder = false
 	}
 
 	if haveBuilder {
 		var err error
+		var file *os.File
 
-		buildBuilderCmd := exec.Command("docker", "build", "-t", containerName+"-builder", "-f", dockerfileBuild, ".")
+		file, err = os.Open(container.DockerfileBuild)
+		if err != nil {
+			log.Error("os.Open", "path", container.DockerfileBuild, "Error", err)
+			return
+		}
+		defer file.Close()
+
+		builderBuildArgs := make([]string, 0)
+
+		scanner := bufio.NewScanner(file)
+		for scanner.Scan() {
+			match := buildArgRe.FindAllStringSubmatch(scanner.Text(), -1)
+			if len(match) == 1 && len(match[0]) == 2 {
+				log.Info("found build arg", "ARG", match[0][1])
+				builderBuildArgs = append(builderBuildArgs, match[0][1])
+			}
+		}
+		if err = scanner.Err(); err != nil {
+			log.Error("scanner", "Error", err)
+			return
+		}
+
+		builderContainerName := container.Name + "-builder"
+
+		buildBuilderCmdArgs := []string{
+			"build",
+			"-t", builderContainerName,
+			"-f", container.DockerfileBuild,
+		}
+		for _, arg := range builderBuildArgs {
+			if value, exists := os.LookupEnv(arg); exists {
+				buildBuilderCmdArgs = append(buildBuilderCmdArgs, "--build-arg", arg+"="+value)
+			} else if container.ValidateBuildArgsExist == nil || *container.ValidateBuildArgsExist == true {
+				log.Error("build arg " + arg + " missing from environment")
+				return
+			}
+		}
+		buildBuilderCmdArgs = append(buildBuilderCmdArgs, ".")
+		log.Debug("docker build", "Args", buildBuilderCmdArgs[1:])
+
+		defer exec.Command("docker", "rmi", "-f", builderContainerName).Run()
+
+		buildBuilderCmd := exec.Command("docker", buildBuilderCmdArgs...)
 		buildBuilderCmd.Stdout = os.Stdout
 		buildBuilderCmd.Stderr = os.Stderr
 		err = buildBuilderCmd.Run()
@@ -202,7 +249,7 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 			return
 		}
 
-		runCmd := exec.Command("docker", "run", "--rm", containerName+"-builder")
+		runCmd := exec.Command("docker", "run", "--rm", builderContainerName)
 
 		runCmdStdoutPipe, err := runCmd.StdoutPipe()
 		if err != nil {
@@ -211,8 +258,8 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 		}
 
 		buildCmd := exec.Command("docker", "build",
-			"-t", containerName,
-			"-f", dockerfile,
+			"-t", container.Name,
+			"-f", container.Dockerfile,
 			"-")
 		buildCmd.Stdin = runCmdStdoutPipe
 		buildCmd.Stdout = os.Stdout
@@ -226,7 +273,7 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 
 		err = buildCmd.Start()
 		if err != nil {
-			log.Error("build Dockerfile", "Error", err)
+			log.Error("docker build", "Error", err)
 			return
 		}
 
@@ -234,8 +281,8 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 		buildCmd.Wait()
 	} else {
 		buildCmd := exec.Command("docker", "build",
-			"-t", containerName,
-			"-f", dockerfile,
+			"-t", container.Name,
+			"-f", container.Dockerfile,
 			".")
 		buildCmd.Stdout = os.Stdout
 		buildCmd.Stderr = os.Stderr
@@ -258,7 +305,7 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 		dockerSaveLock.Lock()
 		defer dockerSaveLock.Unlock()
 
-		saveCmd := exec.Command("docker", "save", "-o", imagePath, containerName)
+		saveCmd := exec.Command("docker", "save", "-o", imagePath, container.Name)
 		saveCmd.Stdout = os.Stdout
 		saveCmd.Stderr = os.Stderr
 		err = saveCmd.Run()
@@ -271,7 +318,7 @@ func buildContainer(log log15.Logger, containerName, dockerfile, dockerfileBuild
 
 		log.Info("docker push")
 
-		pushCmd := exec.Command("docker", "push", containerName)
+		pushCmd := exec.Command("docker", "push", container.Name)
 		pushCmd.Stdout = os.Stdout
 		pushCmd.Stderr = os.Stderr
 		err := pushCmd.Run()